Added API calls
main.js will be the API for managing an army.

Added getArmy to return the Army

Added getModelList to return the library of models that can be added to a Unit.

Added createUnit which takes in an array of model names which will be converted into a Unit and added to the Army